Fixed a build bug caused by the editor's fake display.
In the built game, the wood sword sprite appeared in a layer higher than the player's, while in the editor, it was displayed the way it was supposed to. The built game showed me a different result from what it was promised to me in the editor, justifying this after-jam update.